Updating Fork

+++ app/Http/Controllers/DietPlanController.php
@@ -1,18 +1,20 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Http\Request;
 use App\DietPlan;
 
-class DietPlanController extends Controller {
+class DietPlanController extends Controller
+{
 
-    public function __construct(){
+    public function __construct()
+    {
         $this->middleware('auth', ['except' => ['index', 'show']]);
     }
 
     /**
      * Show the form for creating a new resource.
-     *
      * @return \Illuminate\Http\Response
      */
     public function create()
@@ -22,15 +24,17 @@ class DietPlanController extends Controller {
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $this->validate($request, [
-            'calories_day' => 'required',
+            'weight' => 'required',
+            'date' => 'required',
         ]);
+
+        $this->calculatePlan($request);
 
         // Create DietPlan
         $dietPlan = new DietPlan();
@@ -43,51 +47,59 @@ class DietPlanController extends Controller {
     }
 
     /**
+     * @param \Illuminate\Http\Request $request
+     */
+    protected function calculatePlan(Request $request)
+    {
+        // TODO handle diet plan calculation
+    }
+
+    /**
      * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
+     * @param \App\Http\Controllers\int $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit(int $id)
     {
         $dietPlan = DietPlan::find($id);
         return view('editDietPlan')->with('dietPlan', $dietPlan);
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Http\Controllers\int $id
+     * @return \Illuminate\Http\RedirectResponse
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id)
     {
         $this->validate($request, [
-            'calories_day' => 'required',
+            'weight' => 'required',
+            'date' => 'required',
         ]);
 
-        // Create DietPlan
-        $dietPlan = new DietPlan();
-        $dietPlan->calories_day = $request->input('calories_day');
-        $dietPlan->user_id = auth()->user()->id;
+        $this->calculatePlan($request);
 
+        $dietPlan = DietPlan::find($id);
+        $dietPlan->calories_day = $request->input('calories_day');
         $dietPlan->save();
 
         return redirect('/dashboard')->with('success', 'Diet Plan Updated');
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param \App\Http\Controllers\int $id
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Exception
      */
-    public function destroy($id)
+    public function destroy(int $id)
     {
         $dietPlan = DietPlan::find($id);
-        $dietPlan->delete();
-
+        if (!empty($dietPlan)) {
+            $dietPlan->delete();
+        }
+        else {
+            // todo throw exception
+        }
         return redirect('/dashboard')->with('success', 'Diet Plan Removed');
     }
 }
